Refactor auth-routes, move user creation method into UserController
